Restrict WSGI process lifetime
As a temporary workaround for growing memory usage of the WSGI
processes the processes are configured to respawn after serving 100
requests.  This buys us time to find the source of the memory issues.
The picture is clouded by LRU caching within the processes.

Initial investigations show that most new python-level objects are
related to the request processing.  It isn't clear whether these are
being freed after request completion or not.  It is difficult to track
some of the new objects as they are of generic types (e.g. dict) which
makes them harder to isolate and analyse.