Remove social page
Neil suggested removing the page because it "doesn’t really have any
content and the reference to GroupSpaces is out of date while the
Facebook page seems to be pretty dormant. Basically everything happens
through Spond or WhatsApp, and is internal-facing".

THe page may come back to I have commented out some of the code rather
than removing it; this is the same process we followed when
temporarily disabling the committee page.