Use default venue when importing away fixtures without a venue
When importing a CSV file of fixtures, all away fixtures without a
specific venue are placed at the default venue for the opposition
club.

Jan has put a lot of work into compiling the fixture spreadsheet.
Before this change, the import script recorded fixtures marked as
"away" (or empty) as "TBD" venue.  Jan does fill in the venue if it is
not the default for the opposition club, but it does not make sense
for him to waste time filling in default venues for about half the
games this season.