Remove duplicate "the Club Secretary"
It makes sense to set SecretaryName to be "Club Secretary, XXX"
because it makes it clear this is a role while humanising it with the
current occupant.  Most of the instances of SecretaryName are part of
clubinfo_email tags and we want both "Club Secretary" and "XXX" to be
part of the link.

In this one instance we have the text "...the Club Secretary, {%
clubinfo_email SecretaryName %}" which duplicates the role name.